requirements.txt
absl-py==1.4.0
cachetools==4.2.4
certifi==2025.1.31
charset-normalizer==2.0.12
cycler==0.11.0
Cython==3.0.12
dataclasses==0.8
easydict==1.13
efficientnet-pytorch==0.4.0
future==1.0.0
google-auth==2.22.0
google-auth-oauthlib==0.4.6
grpcio==1.48.2
idna==3.10
importlib-metadata==4.8.3
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
networkx==1.5
numpy==1.19.5
oauthlib==3.2.2
opencv-python==4.0.1.24
Pillow==8.4.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.27.1
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.5.4
six==1.17.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==1.0
typing_extensions==4.1.1
urllib3==1.26.20
Werkzeug==2.0.3
zipp==3.6.0
